Remove React testing setup
