+++ README.md
@@ -3,3 +3,9 @@
 A guessing game built with React and bundled using Vite.
 
 Try it now: https://ambrozy.github.io/iguessgame/
+
+## Running tests
+
+```bash
+npm test
+```
