+++ package.json
@@ -6,14 +6,18 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "jest"
   },
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.27.2",
     "@vitejs/plugin-react": "^4.0.4",
+    "babel-jest": "^30.0.0-beta.3",
+    "jest": "^29.7.0",
     "vite": "^4.4.8"
   }
 }
